feat(python): test, watch, coverage, debug tools

+++ src/conjuring/spells/py.py
@@ -1,3 +1,5 @@
+from pathlib import Path
+
 from invoke import Context, task
 
 from conjuring.grimoire import print_error, run_command, run_lines, run_with_fzf
@@ -36,6 +38,17 @@ class Poetry:
     def __init__(self, context: Context) -> None:
         self.context = context
 
+    def used_in_project(self, display_error=True) -> bool:
+        """Check if Poetry is being used."""
+        used = int(
+            run_command(
+                self.context, f"grep tool.poetry {PYPROJECT_TOML} 2>/dev/null | wc -c", hide=True, warn=True
+            ).stdout.strip()
+        )
+        if not used and display_error:
+            print_error("This task only works with Poetry projects (so far).")
+        return bool(used)
+
     def parse_python_version(self, venv: str):
         """For now, assuming we only have Poetry venvs."""
         return venv.split(" ")[0].split("-py")[1]
@@ -67,6 +80,9 @@ def editable(c, inject=""):
 
     It won't be needed anymore when https://github.com/python-poetry/poetry-core/pull/182 is merged.
     """
+    if not Poetry(c).used_in_project():
+        return
+
     chosen_repo = ""
     if inject:
         # Ask for the repo before doing anything else... to fail fast if no repo is chosen
@@ -94,6 +110,9 @@ def install(c, version="", force=False, delete_all=False):
     """Install a Python virtual environment. For now, only works with Poetry."""
     venv_list = run_lines(c, "poetry env list", hide=False)
     poetry = Poetry(c)
+    if not poetry.used_in_project():
+        return
+
     if delete_all:
         for venv in venv_list:
             poetry.remove_venv(poetry.parse_python_version(venv))
@@ -109,3 +128,53 @@ def install(c, version="", force=False, delete_all=False):
     poetry.use_venv(version)
 
     c.run("poetry install")
+
+
+@task(help={"watch": "Watch for changes and re-run affected tests. Install pytest-watch and pytest-testmon first."})
+def test(c, watch=False):
+    """Run tests with pytest."""
+    if not Poetry(c).used_in_project():
+        return
+
+    command = 'ptw --runner "pytest --testmon"' if watch else "pytest -v"
+    run_command(c, "poetry run", command)
+
+
+@task(help={"show_all": "Show all lines, even if they are covered"})
+def coverage(c, show_all=False):
+    """Run tests with pytest and coverage."""
+    if not Poetry(c).used_in_project():
+        return
+
+    options = [f"--cov={source}" for source in ["src", Path.cwd().name, "app"] if Path(source).exists()]
+
+    skip_option = "" if show_all else ":skip-covered"
+    options.append(f"--cov-report=term-missing{skip_option}")
+
+    run_command(c, "poetry run pytest -v", *options)
+
+
+@task(
+    help={
+        "all_": "Install all debug tools",
+        "ipython": "Install https://pypi.org/project/ipython/",
+        "ipdb": "Install https://pypi.org/project/ipdb/",
+        "pudb": "Install https://pypi.org/project/pudb/",
+        "icecream": "Install https://pypi.org/project/icecream/",
+        "devtools": "Install https://pypi.org/project/devtools/",
+    }
+)
+def debug_tools(c, all_=False, ipython=False, ipdb=False, pudb=False, icecream=False, devtools=False):
+    """Install debug tools."""
+    if not Poetry(c).used_in_project():
+        return
+
+    tools = [
+        "pip",
+        "ipython" if ipython or all_ else "",
+        "ipdb" if ipdb or all_ else "",
+        "pudb" if pudb or all_ else "",
+        "icecream" if icecream or all_ else "",
+        "devtools[pygments]" if devtools or all_ else "",
+    ]
+    run_command(c, "poetry run pip install --upgrade", *tools)
